requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==4.0.0
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
convertdate==2.4.0
cryptography==40.0.1
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
EasyProcess==1.1
entrypoint2==1.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.99.1
fastjsonschema==2.16.3
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.39.3
frozendict==2.3.7
frozenlist==1.4.0
fsspec==2023.9.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.17.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
hijri-converter==2.2.4
holidays==0.22
html5lib==1.1
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.17.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.6
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
langchain==0.0.287
langsmith==0.0.36
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.0.4
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.28.0
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.7.0
plotly==5.14.1
posthog==3.0.2
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.0
pydicom==2.3.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyMeeus==0.5.12
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.1
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
safetensors==0.3.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1+cu117
torchvision==0.15.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.23.1
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.9.2
yfinance==0.2.16
zipp==3.16.2
